requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
dotenv==0.9.9
einops==0.8.1
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
hydra-core==1.3.2
idna==3.10
jaxtyping==0.3.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lightning==2.5.1
lightning-utilities==0.14.2
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
networkx==3.3
numpy==2.2.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.24.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.1
threadpoolctl==3.6.0
torchmetrics==1.7.0
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wadler_lindig==0.1.4
wandb==0.19.8
yarl==1.18.3
fastdtw
h5py
pyarrow
fastparquet